requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-debug-toolbar==2.2
dj-database-url==0.5.0
Pillow==7.1.2
